Refactor execute code
